+++ OUDS/Core/Themes/Wireframe/Sources/Values/ComponentTokens/WireframeTheme+LinkComponentTokens.swift
@@ -29,13 +29,13 @@ extension WireframeThemeLinkComponentTokensProvider: LinkComponentTokens {
     /// Has for value token colors.colorActionFocus
     @objc public final var linkColorContentFocus: MultipleColorSemanticTokens { colors.colorActionFocus }
     /// Has for value token colors.colorActionEnabled
-    @objc public final var linkColorArrowEnabled: MultipleColorSemanticTokens { colors.colorActionEnabled }
+    @objc public final var linkColorChevronEnabled: MultipleColorSemanticTokens { colors.colorActionEnabled }
     /// Has for value token colors.colorActionHover
-    @objc public final var linkColorArrowHover: MultipleColorSemanticTokens { colors.colorActionHover }
+    @objc public final var linkColorChevronHover: MultipleColorSemanticTokens { colors.colorActionHover }
     /// Has for value token colors.colorActionPressed
-    @objc public final var linkColorArrowPressed: MultipleColorSemanticTokens { colors.colorActionPressed }
+    @objc public final var linkColorChevronPressed: MultipleColorSemanticTokens { colors.colorActionPressed }
     /// Has for value token colors.colorActionFocus
-    @objc public final var linkColorArrowFocus: MultipleColorSemanticTokens { colors.colorActionFocus }
+    @objc public final var linkColorChevronFocus: MultipleColorSemanticTokens { colors.colorActionFocus }
     /// Has for value token spaces.spacePaddingInlineNone
     @objc public final var linkSpacePaddingInline: SpaceSemanticToken { spaces.spacePaddingInlineNone }
     /// Has for value token spaces.spacePaddingBlockMd